# apps/kernel/docs/openapi.yaml
openapi: 3.0.3
info:
  title: AI-BOS Kernel API
  version: 1.0.0
  description: |
    Control Plane API for AI-BOS Finance platform.
    
    ## Architecture
    The Kernel is the Control Plane that orchestrates:
    - **Authentication:** JWT-based identity verification
    - **Authorization:** RBAC permission enforcement
    - **Gateway:** Reverse proxy routing to Cells
    - **Audit:** Immutable event logging with correlation IDs
    
    ## Authentication
    Most endpoints require JWT Bearer token (obtained via `/api/kernel/iam/login`).
    
    ## Bootstrap
    First user creation requires `x-kernel-bootstrap-key` header (see Cell Integration Guide).
    
    ## Tenant Authority
    - **Before login:** `x-tenant-id` header selects tenant context
    - **After login:** JWT `tid` claim is authoritative; header is ignored
    
    ## Data Persistence
    All state is persisted to PostgreSQL (survives restarts).
    
    ## Reference Cell
    The `cell-payment-hub` demonstrates cell-based resilience:
    - Health endpoints: `/ping`, `/health`
    - Finance endpoints: `/payments/process`, `/payments/status/:id`
    - Chaos endpoints: `/chaos/fail/:cell`, `/chaos/recover/:cell`
servers:
  - url: http://localhost:3001
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
...
  headers:
    XTenantId:
      schema:
        type: string
      description: Tenant ID (required for control-plane calls and public gateway calls)
      required: true
...
  schemas:
    Error:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, VALIDATION_ERROR, NOT_FOUND, INTERNAL_ERROR, BOOTSTRAP_DENIED]
            message:
              type: string
            details:
              type: array
              items:
                type: object
        correlation_id:
          type: string
...
paths:
...
  /api/kernel/iam/login:
    post:
      summary: Login
      description: Authenticates user and returns JWT access token
      operationId: login
      parameters:
        - $ref: '#/components/headers/XTenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      access_token:
                        type: string
                      token_type:
                        type: string
                        example: Bearer
                      expires_in:
                        type: integer
                  correlation_id:
                    type: string
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/kernel/iam/roles:
    post:
      summary: Create role
      description: Creates a new role (requires RBAC permission `kernel.iam.role.create`)
      operationId: createRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/headers/XTenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Role created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  tenant_id:
                    type: string
                  name:
                    type: string
                  created_at:
                    type: string
                    format: date-time
        '403':
          description: Forbidden (insufficient permissions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: List roles
      description: Lists all roles for the tenant (requires JWT)
      operationId: listRoles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/headers/XTenantId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
...
  /api/kernel/events/publish:
    post:
      summary: Publish event (SECURED - Build 3.4)
      description: |
        Publishes an event to the Event Bus.
        
        **Security (Build 3.4):**
        - Requires valid JWT (Bearer token)
        - Requires `kernel.event.publish` permission
        - `tenant_id` is derived from JWT (not request body)
        - `actor_id` is derived from JWT (not request body)
        - Cross-tenant injection attempts are blocked and audited
        
        **Body tenant_id/actor_id:**
        If provided in request body, they are validated against JWT claims.
        Mismatch results in 400 VALIDATION_ERROR and audit log entry.
      operationId: publishEvent
      tags:
        - Events
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_name, payload]
              properties:
                event_name:
                  type: string
                  minLength: 1
                  maxLength: 128
                  example: canon.demo.ping
                  description: Event type identifier
                source:
                  type: string
                  enum: [kernel, canon, molecule, cell]
                  default: canon
                  description: Event source type
                tenant_id:
                  type: string
                  format: uuid
                  description: Optional - MUST match JWT tenant_id if provided (validated, not used)
                actor_id:
                  type: string
                  format: uuid
                  description: Optional - MUST match JWT user_id if provided (validated, not used)
                correlation_id:
                  type: string
                  format: uuid
                  description: Optional correlation ID (generated if missing)
                timestamp:
                  type: string
                  format: date-time
                  description: Optional timestamp (generated if missing)
                payload:
                  type: object
                  description: Event payload (any JSON object)
      responses:
        '201':
          description: Event published successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  event_id:
                    type: string
                    format: uuid
                  correlation_id:
                    type: string
                    format: uuid
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Validation error (invalid body or cross-tenant mismatch)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (missing or invalid JWT)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (missing kernel.event.publish permission)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
                  - type: object
                    properties:
                      error:
                        type: object
                        properties:
                          missing_permissions:
                            type: array
                            items:
                              type: string
                            example: ["kernel.event.publish"]
        '500':
          description: Internal error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...